refactor(CustomCrudDaoTest, PageTest): move method
- The method testPage() is moved from CustomCrudDaoTest to PageTest to
  put the test in its correct place and also make the classes more
  cohesive.
- No test has been deleted.
